Implement duqtools setup
